requirements.txt
twilio==6.10.1
requests==2.18.4
beautifulsoup4==4.6.0
python-dotenv==0.7.1
skimage==0.0
